requirements.txt
CairoSVG==2.7.1
coral-pytorch==1.4.0
duckdb==0.9.1
fastcore==1.5.29
graphviz==0.20.1
hydra-core==1.3.2
imbalanced-learn==0.11.0
matplotlib==3.8.0
numpy==1.23.2
numba==0.56.4
pandas==2.1.2
pyarrow==11.0.0
pytorch-lightning==1.9.5
optuna==3.3.0
scikit-learn==1.3.1
scikit-posthocs==0.7.0
scipy==1.10.0
sf-hamilton==1.38.0
xgboost==1.7.6

--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
